requirements.txt
Django==1.9.1
djangorestframework==3.3.3
psycopg2==2.6.1
